Payments | Rename all files with prefix 'transaction'    (#161)
Edited files with prefix transaction to item_purchase
