+++ backend/payments/apps/payment_accounts/views.py
@@ -1,5 +1,4 @@
 import rollbar
-
 from rest_framework import status
 from rest_framework.generics import CreateAPIView
 from rest_framework.response import Response
