+++ backend/payments/apps/urls.py
@@ -2,6 +2,6 @@ from django.urls import include, path
 
 urlpatterns = [
     path('payment_accounts/', include('apps.payment_accounts.urls')),
-    path('transactions/', include('apps.transactions.urls')),
+    path('item_purchases/', include('apps.item_purchases.urls')),
     path('external_payments/', include('apps.external_payments.urls')),
 ]
